backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
numpy==1.24.4
scikit-learn==1.3.2
matplotlib==3.8.2
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
aiosqlite==0.19.0
transformers==4.36.0
requests==2.31.0
python-dotenv==1.0.0
